+++ api/src/main/java/io/jsonwebtoken/SigningKeyResolverAdapter.java
@@ -35,14 +35,14 @@ import java.security.Key;
  * {@link SigningKeyResolver} interface that allows subclasses to process only the type of JWS body that
  * is known/expected for a particular case.</p>
  *
- * <p>The {@link #resolveSigningKey(JwsHeader, Claims)} and {@link #resolveSigningKey(JwsHeader, String)} method
+ * <p>The {@link #resolveSigningKey(JwsHeader, Claims)} and {@link #resolveSigningKey(JwsHeader, byte[])} method
  * implementations delegate to the
- * {@link #resolveSigningKeyBytes(JwsHeader, Claims)} and {@link #resolveSigningKeyBytes(JwsHeader, String)} methods
+ * {@link #resolveSigningKeyBytes(JwsHeader, Claims)} and {@link #resolveSigningKeyBytes(JwsHeader, byte[])} methods
  * respectively.  The latter two methods simply throw exceptions:  they represent scenarios expected by
  * calling code in known situations, and it is expected that you override the implementation in those known situations;
  * non-overridden *KeyBytes methods indicates that the JWS input was unexpected.</p>
  *
- * <p>If either {@link #resolveSigningKey(JwsHeader, String)} or {@link #resolveSigningKey(JwsHeader, Claims)}
+ * <p>If either {@link #resolveSigningKey(JwsHeader, byte[])} or {@link #resolveSigningKey(JwsHeader, Claims)}
  * are not overridden, one (or both) of the *KeyBytes variants must be overridden depending on your expected
  * use case.  You do not have to override any method that does not represent an expected condition.</p>
  *
@@ -52,6 +52,7 @@ import java.security.Key;
  * @deprecated since JJWT_RELEASE_VERSION. Use {@link LocatorAdapter LocatorAdapter} with
  * {@link JwtParserBuilder#setKeyLocator(Locator)}
  */
+@SuppressWarnings("DeprecatedIsStillUsed")
 @Deprecated
 public class SigningKeyResolverAdapter implements SigningKeyResolver {
 
@@ -74,13 +75,13 @@ public class SigningKeyResolverAdapter implements SigningKeyResolver {
     }
 
     @Override
-    public Key resolveSigningKey(JwsHeader header, String plaintext) {
+    public Key resolveSigningKey(JwsHeader header, byte[] payload) {
         SignatureAlgorithm alg = SignatureAlgorithm.forName(header.getAlgorithm());
-        Assert.isTrue(alg.isHmac(), "The default resolveSigningKey(JwsHeader, String) implementation cannot " +
+        Assert.isTrue(alg.isHmac(), "The default resolveSigningKey(JwsHeader, byte[]) implementation cannot " +
                 "be used for asymmetric key algorithms (RSA, Elliptic Curve).  " +
-                "Override the resolveSigningKey(JwsHeader, String) method instead and return a " +
+                "Override the resolveSigningKey(JwsHeader, byte[]) method instead and return a " +
                 "Key instance appropriate for the " + alg.name() + " algorithm.");
-        byte[] keyBytes = resolveSigningKeyBytes(header, plaintext);
+        byte[] keyBytes = resolveSigningKeyBytes(header, payload);
         return new SecretKeySpec(keyBytes, alg.getJcaName());
     }
 
@@ -104,18 +105,19 @@ public class SigningKeyResolverAdapter implements SigningKeyResolver {
     }
 
     /**
-     * Convenience method invoked by {@link #resolveSigningKey(JwsHeader, String)} that obtains the necessary signing
-     * key bytes.  This implementation simply throws an exception: if the JWS parsed is a plaintext JWS, you must
-     * override this method or the {@link #resolveSigningKey(JwsHeader, String)} method instead.
+     * Convenience method invoked by {@link #resolveSigningKey(JwsHeader, byte[])} that obtains the necessary signing
+     * key bytes.  This implementation simply throws an exception: if the JWS parsed is a payload JWS, you must
+     * override this method or the {@link #resolveSigningKey(JwsHeader, byte[])} method instead.
      *
      * @param header  the parsed {@link JwsHeader}
-     * @param payload the parsed String plaintext payload
+     * @param payload the byte array payload
      * @return the signing key bytes to use to verify the JWS signature.
      */
-    public byte[] resolveSigningKeyBytes(JwsHeader header, String payload) {
+    @SuppressWarnings("unused")
+    public byte[] resolveSigningKeyBytes(JwsHeader header, byte[] payload) {
         throw new UnsupportedJwtException("The specified SigningKeyResolver implementation does not support " +
-                "plaintext JWS signing key resolution.  Consider overriding either the " +
-                "resolveSigningKey(JwsHeader, String) method or, for HMAC algorithms, the " +
-                "resolveSigningKeyBytes(JwsHeader, String) method.");
+                "payload JWS signing key resolution.  Consider overriding either the " +
+                "resolveSigningKey(JwsHeader, byte[]) method or, for HMAC algorithms, the " +
+                "resolveSigningKeyBytes(JwsHeader, byte[]) method.");
     }
 }