+++ impl/src/main/java/io/jsonwebtoken/impl/security/LocatingKeyResolver.java
@@ -23,7 +23,7 @@ public class LocatingKeyResolver implements SigningKeyResolver {
     }
 
     @Override
-    public Key resolveSigningKey(JwsHeader header, String plaintext) {
+    public Key resolveSigningKey(JwsHeader header, byte[] payload) {
         return this.locator.locate(header);
     }
 }