+++ CHANGELOG.md
@@ -9,35 +9,58 @@
   signature algorithms could not do so until now.  The new interface now allows anyone to plug in and support custom
   algorithms with JJWT as desired.
 
+
 * Similarly, as the `io.jsonwebtoken.security.Keys#secretKeyFor` and `io.jsonwebtoken.security.Keys#keyPairFor` methods 
   accepted the now-deprecated `io.jsonwebtoken.SignatureAlgorithm` enum, they have also been deprecated in favor of 
   calling new `keyBuilder()` or `keyPairBuilder()` methods on `SignatureAlgorithm` instances directly. The builders 
   allow for customization of the JCA `Provider` and `SecureRandom` during Key or KeyPair generation if desired, whereas
   the old enum-based static utility methods did not.
 
+
 * `io.jsonwebtoken.CompressionCodec` now inherits `io.jsonwebtoken.Identifiable` and `getId()` is preferred over
   the now-deprecated `getAlgorithmName()` method.  This was to guarantee API congruence with all other JWT-identifiable
   algorithm names that can be set as a header value.
 
-#### Backwards Compatibility Warnings and Potential Breaking Changes
+#### Backwards Compatibility Warnings and Breaking Changes
 
-* `JwtBuilder` has a new `setPayload(byte[])` method that allows any kind of content within a JWT, not just Strings. 
-  The existing `setPayload(String)` method implementation has been changed to delegate to this new `setPayload(byte[])`
-  method with the argument's UTF-8 bytes, for example `setPayload(aString.getBytes(StandardCharsets.UTF_8))`.
+* JWTs that do not contain JSON Claims now have a body type of `byte[]` instead of `String` (that is, `Jws<byte[]>`
+  instead of `Jws<String>`).  This is because JWTs, 
+  along with the `cty` (Content Type) header, are capable of handling _any_ type of payload, not just Strings. 
+  The previous JJWT releases never accounted for this, and now the API accurately reflects the JWT RFC specification 
+  payload capabilities.  This change has impacted the following JJWT APIs:
+
+  * The `JwtParser`'s `Jwt<Header, String> parsePlaintextJwt(String plaintextJwt)` and 
+    `Jws<String> parsePlaintextJws(String plaintextJws)` methods have been changed to
+    `Jwt<Header, byte[]> parsePlaintextJwt(String plaintextJwt)` and
+    `Jws<byte[]> parsePlaintextJws(String plaintextJws)` respectively.
+
+  * `JwtHandler`'s `onPlaintextJwt` and `onPlaintextJws` method argument types have been changed from `String` to 
+    `byte[]`.
+
+  * `io.jsonwebtoken.JwtHandlerAdapter` has been changed to reflect the above-mentioned `String`-to-`byte[]` argument 
+    changes, as well adding the `abstract` modifier.  This class was never intended
+    to be instantiated directly, and is provided for subclassing only.  The missing modifier has been added to ensure
+    the class is used as it had always been intended.
+
+  * `io.jsonwebtoken.SigningKeyResolver`'s `resolveSigningKey(JwsHeader, String)` method has been changed to
+    `resolveSigningKey(JwsHeader, byte[])`.
+
+  * `JwtBuilder` has a new `setPayload(byte[])` method that allows any kind of content within a JWT, not just Strings. 
+    The existing `setPayload(String)` method implementation has been changed to delegate to this new `setPayload(byte[])`
+    method with the argument's UTF-8 bytes, for example `setPayload(aString.getBytes(StandardCharsets.UTF_8))`.
+
 
 * Prior to this release, if there was a serialization problem when serializing the JWT Header, an `IllegalStateException`
   was thrown. If there was a problem when serializing the JWT claims, an `IllegalArgumentException` was
   thrown.  This has been changed up to ensure consistency: any serialization problem with either headers or claims
-  will throw a `io.jsonwebtoken.io.SerializationException`.
+  will now throw a `io.jsonwebtoken.io.SerializationException`.
+
 
 * Parsing of unsecured JWTs (`alg` header of `none`) are now disabled by default as mandated by 
   [RFC 7518, Section 3.6](https://datatracker.ietf.org/doc/html/rfc7518#section-3.6). If you require parsing of
   unsecured JWTs, you may call the `enableUnsecuredJws` method on the `JwtParserBuilder`, but note the security
   implications of doing so as mentioned in that method's JavaDoc before doing so.
 
-* `io.jsonwebtoken.JwtHandlerAdapter` has been changed to add the `abstract` modifier.  This class was never intended
-  to be instantiated directly, and is provided for subclassing only.  The missing modifier has been added to ensure
-  the class is used as it had always been intended.
 
 * `io.jsonwebtokne.gson.io.GsonSerializer` now requires `Gson` instances that have a registered
   `GsonSupplierSerializer` type adapter, for example: